Added 2 methods & changes to file writing
2 methods was added to JSONmaster.cs, ConvertToJson and ConvertFromJson.
Changed the way it used to write text to files. Instead of using a StreamWriter, now it uses File.WriteAllText method.
I used to verify if the save directory exists befores trying to create the directory, but thre's no need to verify it beforehand because the method will only create directories that doesn't exist yet.